Make sure that the global-query server can be built on any platform
Moves the shellformat stuff to it's own package below types, since it
contains LINUX specifics. The global-query server should be platform
agnostic.

Also update the example configs for the global-query server so it can
be run as described in the README
