+++ examples/config/global-query-example-config.yaml
@@ -7,3 +7,6 @@ hosts:
 querier:
   type: api
   max_concurrent: 64
+  config: ./examples/config/global-query-api-client-querier-example-config.yaml
+server:
+  addr: localhost:8145
