+++ pkg/api/spec/openapi.yaml
@@ -1,4 +1,3 @@
-# TODO: generated by ChatGPT: double check!!
 openapi: 3.0.0
 info:
   title: Query API
